fix: update supported dtypes
Update supported dtypes for certain functions, and change the safety
factor for logsigmoid. All tests now pass locally, except problems
caused by tensorflow's sigmoid and a problem with the test for elu
(which seems to feed in the wrong values).

+++ ivy/functional/ivy/activations.py
@@ -724,7 +724,7 @@ def _hardswish_jax_like(
     def hard_sigmoid(x):
         return ivy.relu6(x + 3.0) / 6
 
-    return ivy.multiply(x, hard_sigmoid(x))
+    return ivy.multiply(x, hard_sigmoid(x).astype(x.dtype))
 
 
 @handle_exceptions
